Added FirstBoot module business logic

+++ app/src/main/java/com/emelyanov/rassvet/modules/firstboot/domain/models/SectionDetailsViewState.kt
@@ -6,6 +6,6 @@ sealed class SectionDetailsViewState {
     data class PresentInfo(
         val title: String,
         val description: String,
-        val price: Float
+        val price: Int
     ) : SectionDetailsViewState()
 }
